Initial project setup
Includes transfer command and console wrapper "executable" and a very
basic unit test.

The .env file contains the variable names to be set when executing the
command. This can either be done inside the .env or .env.local file,
or by setting the env variables in a more traditional way.